- some significant updates to the code::blocks project. It now compiles under cygwin, so it should be able to get the project working under linux again! - some minor changes to remove stupid gcc warnings about things like "no blank line at the end of the file) - a minor change to fix a potential problem with the use of the is*() functions and to*() functions for text (like isdigit and toupper). These seem to still be implemented as macros that access arrays on gcc-3, so I mage sure to cast them to unsigned so they will never access elements before the start of the array (since char is signed).

+++ egolib/scancode.c
@@ -50,11 +50,11 @@ void scancode_begin( void )
 		dst = src;
         if ( src < 255 )
         {
-			dst = toupper( src );
+			dst = toupper( (unsigned)src );
 
 			if( dst > 255 )
 			{
-				fprintf( stderr, "%s - toupper() returned an out of range value \'%c\' -> %d\n", __FUNCTION__, src, dst ); 
+				fprintf( stderr, "%s - toupper() returned an out of range value \'%c\' -> %d\n", __FUNCTION__, src, dst );
 			}
         }
